Drop sorts validation and unused helpers cleanup
JIRA: TNT-451
