+++ libs/sdk-ui-ext/src/internal/utils/tests/sort.test.ts
@@ -1,19 +1,15 @@
 // (C) 2019-2022 GoodData Corporation
-import cloneDeep from "lodash/cloneDeep";
 import { BucketNames } from "@gooddata/sdk-ui";
 import {
     createSorts,
     getAttributeSortItem,
     getBucketItemIdentifiers,
     getDefaultTreemapSortFromBuckets,
-    removeInvalidSort,
-    setSortItems,
 } from "../sort";
 import { IExtendedReferencePoint } from "../../interfaces/Visualization";
 import * as referencePointMocks from "../../tests/mocks/referencePointMocks";
 import { DEFAULT_BASE_CHART_UICONFIG } from "../../constants/uiConfig";
-import { SORT_DIR_ASC, SORT_DIR_DESC } from "../../constants/sort";
-import { ATTRIBUTE, FILTERS, METRIC } from "../../constants/bucket";
+import { SORT_DIR_DESC } from "../../constants/sort";
 import {
     insightWithSingleMeasureAndStack,
     insightWithSingleMeasureAndTwoViewBy,
@@ -200,240 +196,6 @@ describe("getAttributeSortItem", () => {
         expect(getAttributeSortItem("a2")).toEqual({
             attributeSortItem: { attributeIdentifier: "a2", direction: "asc" },
         });
-    });
-});
-
-describe("removeInvalidSort", () => {
-    it("should remove sorts with with identifiers that are not included in buckets", () => {
-        const invalidAttributeSort: IAttributeSortItem = {
-            attributeSortItem: {
-                attributeIdentifier: "invalid",
-                direction: SORT_DIR_DESC,
-            },
-        };
-        const invalidMeasureSort1: IMeasureSortItem = {
-            measureSortItem: {
-                direction: "asc",
-                locators: [
-                    {
-                        attributeLocatorItem: {
-                            attributeIdentifier: "invalid",
-                            element: "/gdc/md/PROJECTID/obj/2210/elements?id=1234",
-                        },
-                    },
-                ],
-            },
-        };
-        const invalidMeasureSort2: IMeasureSortItem = {
-            measureSortItem: {
-                direction: "asc",
-                locators: [
-                    {
-                        measureLocatorItem: {
-                            measureIdentifier: "invalid",
-                        },
-                    },
-                ],
-            },
-        };
-        const referencePointWithInvalidSortItems = {
-            ...referencePoint,
-            properties: {
-                sortItems: [
-                    ...referencePoint.properties.sortItems,
-                    invalidAttributeSort,
-                    invalidMeasureSort1,
-                    invalidMeasureSort2,
-                ],
-            },
-        };
-
-        const sanitizedSortItems = removeInvalidSort(referencePointWithInvalidSortItems).properties.sortItems;
-
-        expect(sanitizedSortItems).toEqual(referencePoint.properties.sortItems);
-    });
-});
-
-describe("setSortItems", () => {
-    const emptyReferencePoint: IExtendedReferencePoint = {
-        buckets: [],
-        filters: {
-            localIdentifier: FILTERS,
-            items: [],
-        },
-        uiConfig: {
-            buckets: {},
-        },
-        properties: {
-            sortItems: [],
-        },
-    };
-
-    it("should return unchanged reference if there are sort properties present", () => {
-        const measureSortItem: IMeasureSortItem = {
-            measureSortItem: {
-                direction: "asc",
-                locators: [
-                    {
-                        measureLocatorItem: {
-                            measureIdentifier: "invalid",
-                        },
-                    },
-                ],
-            },
-        };
-
-        const referencePointWithSortItems: IExtendedReferencePoint = cloneDeep({
-            ...emptyReferencePoint,
-            properties: {
-                sortItems: [measureSortItem],
-            },
-        });
-
-        expect(setSortItems(referencePointWithSortItems)).toEqual(referencePointWithSortItems);
-    });
-
-    it("should return unchanged reference if there are no sort properties nor bucket items present", () => {
-        const noSortNoItemsRefPoint: IExtendedReferencePoint = cloneDeep({
-            ...emptyReferencePoint,
-        });
-
-        expect(setSortItems(noSortNoItemsRefPoint)).toEqual(noSortNoItemsRefPoint);
-    });
-
-    it("should add descending sort item based on first measure if there is a measure but no attribute", () => {
-        const onlyMeasureRefPoint: IExtendedReferencePoint = cloneDeep({
-            ...emptyReferencePoint,
-            buckets: [
-                {
-                    localIdentifier: "measures",
-                    items: [
-                        {
-                            type: METRIC,
-                            localIdentifier: "m1",
-                        },
-                        {
-                            type: METRIC,
-                            localIdentifier: "m2",
-                        },
-                    ],
-                },
-            ],
-        });
-
-        expect(setSortItems(onlyMeasureRefPoint).properties).toEqual({
-            sortItems: [
-                {
-                    measureSortItem: {
-                        direction: SORT_DIR_DESC,
-                        locators: [
-                            {
-                                measureLocatorItem: {
-                                    measureIdentifier: "m1",
-                                },
-                            },
-                        ],
-                    },
-                },
-            ],
-        });
-    });
-
-    it("should skip invalid arithmetic measures", () => {
-        const onlyMeasureRefPoint: IExtendedReferencePoint = cloneDeep({
-            ...emptyReferencePoint,
-            buckets: [
-                {
-                    localIdentifier: "measures",
-                    items: [
-                        {
-                            type: METRIC,
-                            localIdentifier: "am1",
-                            operator: "sum",
-                            operandLocalIdentifiers: ["m1", null],
-                        },
-                        {
-                            type: METRIC,
-                            localIdentifier: "am2",
-                            operator: "sum",
-                            operandLocalIdentifiers: ["m1", "m1"],
-                        },
-                        {
-                            type: METRIC,
-                            localIdentifier: "m1",
-                        },
-                    ],
-                },
-            ],
-        });
-
-        expect(setSortItems(onlyMeasureRefPoint).properties).toEqual({
-            sortItems: [
-                {
-                    measureSortItem: {
-                        direction: SORT_DIR_DESC,
-                        locators: [
-                            {
-                                measureLocatorItem: {
-                                    measureIdentifier: "am2",
-                                },
-                            },
-                        ],
-                    },
-                },
-            ],
-        });
-    });
-
-    it("should add ascending sort item based on first attribute if there is an attribute but no measure", () => {
-        const onlyAttributeRefPoint: IExtendedReferencePoint = cloneDeep({
-            ...emptyReferencePoint,
-            buckets: [
-                {
-                    localIdentifier: "measures",
-                    items: [
-                        {
-                            type: ATTRIBUTE,
-                            localIdentifier: "a1",
-                        },
-                    ],
-                },
-            ],
-        });
-
-        expect(setSortItems(onlyAttributeRefPoint).properties).toEqual({
-            sortItems: [
-                {
-                    attributeSortItem: {
-                        attributeIdentifier: "a1",
-                        direction: SORT_DIR_ASC,
-                    },
-                },
-            ],
-        });
-    });
-
-    it("should not add any sort items if there are both measure and attribute", () => {
-        const refPoint: IExtendedReferencePoint = cloneDeep({
-            ...emptyReferencePoint,
-            buckets: [
-                {
-                    localIdentifier: "misc",
-                    items: [
-                        {
-                            type: ATTRIBUTE,
-                            localIdentifier: "a1",
-                        },
-                        {
-                            type: METRIC,
-                            localIdentifier: "m1",
-                        },
-                    ],
-                },
-            ],
-        });
-
-        expect(setSortItems(refPoint)).toBe(refPoint);
     });
 });
